Replace ft_custom_atoi() with ft_secure_atol()
custom_atoi() was bugged and easily misusable,
as there was no way to tell if there is still content
after the numeric part of the string.
There would be in theory, but the internal static function
used an boolean AND instead of a multiplication in a return
meaning the return value of custom_atoi() was not what was
advertised in the doc comment all along.

secure_atol upgrades to long type,
checks and refuses trailing content,
and fixes the aforementioned bug.